requirements.txt
dataclasses
psycopg2
sqlalchemy
chromadb==0.3.27
dataclasses_json
openai
# langchain==0.0.162
langchain==0.0.198 
# next version breaks it!
typing-inspect==0.8.0 
typing_extensions==4.5.0
azure-search-documents
flask